更新
BOT除外処理を追加。

+++ cogs/ara_ng.py
@@ -9,11 +9,12 @@ class ara_ng(commands.Cog):
         
     @commands.Cog.listener()
     async def on_message(self, message):
-      ruizisan = os.path.join(os.path.dirname(__file__), 'ngwords.txt')
-        with open(ruizisan, 'r', encoding='utf-8') as file:
-            bad_words = [bad_word.strip().lower() for bad_word in file.readlines()]
-            if any(bad_word in message.content for bad_word in bad_words):
-                await message.delete()
+      if not message.author.bot:
+          ruizisan = os.path.join(os.path.dirname(__file__), 'ngwords.txt')
+            with open(ruizisan, 'r', encoding='utf-8') as file:
+                bad_words = [bad_word.strip().lower() for bad_word in file.readlines()]
+                if any(bad_word in message.content for bad_word in bad_words):
+                    await message.delete()
 
 
 def setup(bot):
